+++ src/frontend/src/components/SearchHelp/index.js
@@ -1,0 +1,3 @@
+import SearchHelp from './SearchHelp.jsx';
+
+export default SearchHelp;
